Add data_schema to svgfigure
